Remove model downloader component and its usages, move OMZ precision enum
