+++ client/e2e/src/pages/hugging-face-model-download.po.ts
@@ -2,7 +2,7 @@ import { browser, by, element, ElementArrayFinder, ElementFinder, protractor } f
 
 import { TestUtils } from './test-utils';
 import { ModelFile } from './model-file';
-import { OMZModelPrecisionEnum } from '../../../src/app/modules/model-manager/components/model-downloader-table/model-downloader-table.component';
+import { OMZModelPrecisionEnum } from '../../../src/app/modules/model-manager/components/omz-import-ribbon-content/omz-import-ribbon-content.component';
 
 export const filterGroupNames = ['task', 'library', 'model-type', 'language', 'license'];
 type FilterGroupName = typeof filterGroupNames[number];
