--- client/src/app/modules/model-manager/components/model-downloader-table/model-downloader-table.component.html
@@ -1,135 +1,0 @@
-<wb-table-filter-form
-  [columns]="filterableColumns"
-  [columnOptions]="filterableColumnOptions"
-  (applyFilter)="applyFilter($event)"
-  (resetFilter)="resetFilter()"
-></wb-table-filter-form>
-
-<div *ngIf="!isLoading" class="downloads-table-container">
-  <table
-    mat-table
-    [dataSource]="modelsDataSource"
-    id="model-download-table"
-    class="wb-table wb-table-on-gray"
-    matSort
-    multiTemplateDataRows
-    (matSortChange)="sortedColumn = $event"
-  >
-    <ng-container [matColumnDef]="OMZColumnNames.MODEL_NAME">
-      <th mat-header-cell *matHeaderCellDef class="name-column" mat-sort-header>
-        {{ OMZColumnLabels.MODEL_NAME }}
-        <ng-container *ngTemplateOutlet="sortIcon; context: { columnName: OMZColumnNames.MODEL_NAME }"></ng-container>
-      </th>
-      <td id="model-name" mat-cell *matCellDef="let element" class="name-column">
-        <span class="name" data-test-id="model-name">{{ element.name }}</span>
-      </td>
-    </ng-container>
-
-    <ng-container [matColumnDef]="OMZColumnNames.DETAILS">
-      <th mat-header-cell *matHeaderCellDef class="details-column"></th>
-      <td mat-cell *matCellDef="let element" class="details-column">
-        <span
-          [ngClass]="element.name === selectedModel?.name && detailsShown ? 'wb-body-4' : 'wb-body-5'"
-          (click)="selectModelDetails(element); $event.stopPropagation()"
-          data-test-id="details"
-        >
-          {{ OMZColumnLabels.DETAILS }}
-        </span>
-      </td>
-    </ng-container>
-
-    <ng-container [matColumnDef]="OMZColumnNames.PRECISION">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>
-        {{ OMZColumnLabels.PRECISION }}
-        <ng-container *ngTemplateOutlet="sortIcon; context: { columnName: OMZColumnNames.PRECISION }"></ng-container>
-      </th>
-      <td id="model-precision" mat-cell *matCellDef="let element">
-        {{ element.precision }}
-      </td>
-    </ng-container>
-
-    <ng-container [matColumnDef]="OMZColumnNames.FRAMEWORK">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>
-        {{ OMZColumnLabels.FRAMEWORK }}
-        <ng-container *ngTemplateOutlet="sortIcon; context: { columnName: OMZColumnNames.FRAMEWORK }"></ng-container>
-      </th>
-      <td id="model-framework" mat-cell *matCellDef="let element">
-        {{ modelFrameworkNamesMap[element.framework] }}
-      </td>
-    </ng-container>
-
-    <ng-container [matColumnDef]="OMZColumnNames.TASK_TYPE">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>
-        {{ OMZColumnLabels.TASK_TYPE }}
-        <ng-container *ngTemplateOutlet="sortIcon; context: { columnName: OMZColumnNames.TASK_TYPE }"></ng-container>
-      </th>
-      <td id="model-task" mat-cell *matCellDef="let element">
-        {{ ModelTaskTypeToNameMap[element.task_type] }}
-      </td>
-    </ng-container>
-
-    <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-
-    <tr
-      mat-row
-      *matRowDef="let row; columns: displayedColumns"
-      [id]="getModelDownloaderDTOId(row.name, omzModelRowClassPrefix)"
-      #omzModelRow
-      [class.selected]="isRowSelected(row)"
-      (click)="selectedModel = row"
-    ></tr>
-  </table>
-</div>
-
-<div class="loading" *ngIf="isLoading">
-  <wb-spinner></wb-spinner>
-</div>
-
-<div class="model-precision">
-  <wb-config-form-field
-    *ngIf="selectedModel?.framework === ModelFrameworks.OPENVINO"
-    class="hint-first enlarged-control"
-    [field]="openvinoModelPrecisionField"
-    [group]="openvinoModelPrecisionFormGroup"
-  ></wb-config-form-field>
-</div>
-<div *ngIf="isGettingFrameworksAvailabilityFailed; else importNotification" class="error-container">
-  <wb-message-box
-    messageType="error"
-    message="Error on getting available frameworks information"
-    [detailedMessage]="frameworksAvailability.error"
-  >
-  </wb-message-box>
-</div>
-<div class="panel-row panel-row--footer">
-  <wb-button
-    class="enlarged-control"
-    testId="download-btn"
-    type="primary"
-    text="Download and Import"
-    (handleClick)="importModel(selectedModel)"
-    [disabled]="isImportDisabled"
-  >
-  </wb-button>
-  <wb-button class="enlarged-control" testId="cancel-btn" text="Cancel" (handleClick)="cancel.emit()"></wb-button>
-  <wb-info-hint *ngIf="isConnected && !selectedModel" message="Click on table row to select model"></wb-info-hint>
-</div>
-<div *ngIf="!isConnected" class="no-connect-message" data-test-id="no-connect-message">
-  <wb-info-hint class="wb-full-width" [message]="noConnectionMessage" type="attention"></wb-info-hint>
-</div>
-<div class="unavailable-model-message" *ngIf="isUnavailableModel">
-  <wb-info-hint [message]="unavailableOmzModelMessage" type="attention"></wb-info-hint>
-</div>
-
-<ng-template #sortIcon let-columnName="columnName">
-  <ng-container *ngIf="modelsDataSource.data?.length">
-    <wb-table-sort-icon [columnName]="columnName" [sortedColumn]="sortedColumn"></wb-table-sort-icon>
-  </ng-container>
-</ng-template>
-
-<ng-template #importNotification>
-  <div *ngIf="!isImportDisabled && !isSelectedFrameworkConfigured" class="import-notification">
-    <span>Note:</span>
-    <span [textContent]="frameworkAvailabilityNote"></span>
-  </div>
-</ng-template>
