--- client/src/app/modules/model-manager/components/model-downloader-table/model-downloader-table.component.scss
@@ -1,78 +1,0 @@
-@import 'variables';
-
-:host {
-  wb-table-filter-form {
-    display: flex;
-    margin: 30px $container-padding $container-padding;
-  }
-
-  .name-column {
-    width: 1%;
-    white-space: nowrap;
-  }
-
-  .details-column {
-    width: 30%;
-
-    span {
-      color: $plum;
-      cursor: pointer;
-    }
-  }
-
-  .downloads-table {
-    &-header {
-      display: flex;
-      align-items: center;
-      .filter-field {
-        margin-right: $container-padding;
-      }
-    }
-    &-container {
-      max-height: 240px;
-      overflow-y: auto;
-      border: 1px $very-light-pink solid;
-      border-radius: 3px;
-    }
-  }
-
-  .panel-row {
-    display: flex;
-    align-content: center;
-
-    wb-button + wb-button {
-      margin-left: $container-padding;
-    }
-  }
-
-  .loading {
-    padding-top: $container-padding;
-  }
-
-  .filter-field {
-    padding-left: $container-padding;
-    padding-top: $container-padding;
-  }
-
-  .model-precision {
-    padding: $container-padding 0 0;
-
-    ::ng-deep {
-      .field-label {
-        min-width: auto;
-      }
-
-      .mat-form-field-wrapper {
-        padding-bottom: 0;
-      }
-    }
-  }
-}
-
-.panel-row--footer {
-  align-items: center;
-
-  wb-button + wb-info-hint {
-    margin-left: 20px;
-  }
-}
